requirement.txt
cv
numpy
keras.models
keras.layers
keras.optimizers
keras.preprocessing.image
os
